refactor: 合并新增与编辑门店逻辑，优化表单处理
